back-end/requirements.txt
fastapi==0.103.2
sqlmodel==0.0.8
uvicorn==0.24.0        
pydantic==1.10.13
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
anyio==4.0.0           